perf: GPU argmax, GPU PE, O(n log n) LIS, unordered_map, mel float/fusion, timing

+++ src/asr/mel/mel.hpp
@@ -32,7 +32,7 @@ struct Input {
 };
 
 struct Window {
-    std::vector<double> data;
+    std::vector<float> data;
     int length;
 };
